requirements.txt
bleach==6.0.0
bokeh==3.1.1
cdsapi==0.6.1
certifi==2022.12.7
cftime==1.6.2
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorcet==3.0.1
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
fonttools==4.39.4
fsspec==2023.5.0
holoviews==1.16.0
hvplot==0.8.3
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
kiwisolver==1.4.4
linkify-it-py==2.0.2
locket==1.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
MetPy==1.5.0
ncplot==0.3.2
nctoolkit==0.9.2
netCDF4==1.6.3
numpy==1.24.3
packaging==23.1
pandas==2.0.1
panel==1.0.0
param==1.13.0
partd==1.4.0
Pillow==9.5.0
Pint==0.21
platformdirs==3.5.1
pooch==1.7.0
pyct==0.5.0
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3
pyviz-comms==2.2.1
PyYAML==6.0
requests==2.29.0
scipy==1.10.1
six==1.16.0
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.15
webencodings==0.5.1
xarray==2023.4.2
xyzservices==2023.2.0
zipp==3.15.0
